+++ presentacio-src/net/conselldemallorca/helium/presentacio/mvc/PermisController.java
@@ -41,16 +41,16 @@ public class PermisController extends BaseController {
 	private PermisService permisService;
 	private Validator annotationValidator;
 	private Validator additionalValidator;
-	
+
 	private String codi = "";
-	
+
 	@Autowired
 	public PermisController(
 			PermisService permissionService) {
 		this.permisService = permissionService;
 		additionalValidator = new PermisValidator(permisService);
 	}
-	
+
 	@ModelAttribute("command")
 	public PermisCommand populateCommand(@RequestParam(value = "codi", required = false) String codi) {
 		if ((codi != null) && (!codi.equals(""))) {
@@ -66,7 +66,7 @@ public class PermisController extends BaseController {
 		}
 		return new PermisCommand();
 	}
-	
+
 	@RequestMapping(value = "llistat")
 	public String llistat(
 			HttpServletRequest request,
@@ -93,12 +93,12 @@ public class PermisController extends BaseController {
 								getObjectsPerPage(objectsPerPage))));
 		return "rol/llistat";
 	}
-	
+
 	@RequestMapping(value = "form", method = RequestMethod.GET)
 	public String formGet() {
 		return "rol/form";
 	}
-	
+
 	@RequestMapping(value = "form", method = RequestMethod.POST)
 	public String formPost(
 			HttpServletRequest request,
@@ -132,20 +132,25 @@ public class PermisController extends BaseController {
 			return "redirect:/rol/llistat.html";
 		}
 	}
-	
+
 	@RequestMapping(value = "delete")
 	public String deleteAction(
 			HttpServletRequest request,
 			@RequestParam(value = "codi", required = true) String codi) {
 		if ((!codi.equalsIgnoreCase("HEL_ADMIN")) && (!codi.equalsIgnoreCase("HEL_USER"))) {
-			permisService.deletePermis(codi);
-			missatgeInfo(request, "El rol s'ha esborrat correctament");
+			Permis permis = permisService.getPermisByCodi(codi);
+			if (permis.getUsuaris().size() > 0) {
+				missatgeError(request, "Hi ha usuaris que empren aquest rol");
+			} else {
+				permisService.deletePermis(codi);
+				missatgeInfo(request, "El rol s'ha esborrat correctament");
+			}
 		} else {
 			missatgeError(request, "El rol d'administrador i el rol d'usuari no es poden esborrar");
 		}
 		return "redirect:/rol/llistat.html";
 	}
-	
+
 	@InitBinder
 	public void initBinder(WebDataBinder binder) {
 		binder.registerCustomEditor(
@@ -158,7 +163,7 @@ public class PermisController extends BaseController {
 				Permis.class,
 				new PermisTypeEditor());
 	}
-	
+
 	@Resource(name = "annotationValidator")
 	public void setAnnotationValidator(Validator annotationValidator) {
 		this.annotationValidator = annotationValidator;
@@ -186,6 +191,7 @@ public class PermisController extends BaseController {
 			}
 		}
 	}
-	
+
 	private static final Log logger = LogFactory.getLog(PersonaController.class);
+
 }
